+++ CHANGES.rst
@@ -8,10 +8,11 @@ LATEST
 - Added ControlCodeParser to create human readable text from raw text with control codes in it.
 - Added readonly logic for Text and TextBox.
 - Added ability to enable/disable widgets by column in layouts.
+- Added left/right/up/down navigation to nearest widget.
 - Fixed exception on reinstating NoneType signal handler.
 - Fixed float/int issue with recent builds of pywin32.
 - Fixed issue where setting options changed the selected value (even if it was still present).
-- Fixed erroneuos trigger of on_load for all Frames at start of day.
+- Fixed erroneous trigger of on_load for all Frames at start of day.
 - Fixed bug where Frames passed on events that they already handled.
 - Fixed bug: Restore current theme on screen resize.